las: Fix double->int->double forward-backward epsilon comparison failure
See:
https://github.com/potree/PotreeConverter/issues/535#issuecomment-2692336475

I found some inputs for which

    ux = ux:  1.0000000000000004
    1.0 + eps = 1.0000000000000002

so `ux <= 1.0 + eps` was still out of bounds,
triggering the error unjustifiedly when the las bounding box was correct.

I believe that the double->int32->double conversion added in commit

    5307150e - fix crash if scale too precise  (#421, #426). Still needs testing.

which adds floating-point inaccuracy, thus causing this;
removing it again, `ux` becomes an exact `1.0`, making the problem go away.

A fair question would be:
    After `ux <= 1.0 + eps`, shouldn't we still need to clamp `ux = min(ux, 1.0)`?
Answer:
    No, because the code already does
        int64_t ix = int64_t(std::min(dGridSize * ux, dGridSize - 1.0));
    when converting to a grid index, so that `min()` clamps it
    to not overflow the grid already.